+++ lib/helpers.dart
@@ -9,6 +9,7 @@
 
 import "dart:io";
 import "package:currency_formatter/currency_formatter.dart";
+
 import "package:one_context/one_context.dart";
 import "package:url_launcher/url_launcher.dart";
 import "package:audioplayers/audioplayers.dart";
@@ -133,3 +134,4 @@ String renderCurrency(double? amount, String currency, {int decimals = 2}) {
 
   return value;
 }
+